+++ Packages/com.unity.inputsystem/InputSystem/Plugins/PlayerInput/PlayerInputManager.cs
@@ -243,14 +243,27 @@ namespace UnityEngine.InputSystem
             }
         }
 
+        /// <summary>
+        /// Reference to the prefab that the manager will instantiate when players join.
+        /// </summary>
+        /// <value>Prefab to instantiate for new players.</value>
         public GameObject playerPrefab
         {
             get => m_PlayerPrefab;
             set => m_PlayerPrefab = value;
         }
 
+        /// <summary>
+        /// Singleton instance of the manager.
+        /// </summary>
+        /// <value>Singleton instance or null.</value>
         public static PlayerInputManager instance { get; private set; }
 
+        /// <summary>
+        /// Allow players to join the game based on <see cref="joinBehavior"/>.
+        /// </summary>
+        /// <seealso cref="DisableJoining"/>
+        /// <seealso cref="joiningEnabled"/>
         public void EnableJoining()
         {
             switch (m_JoinBehavior)
@@ -291,6 +304,11 @@ namespace UnityEngine.InputSystem
             m_AllowJoining = true;
         }
 
+        /// <summary>
+        /// Inhibit players from joining the game.
+        /// </summary>
+        /// <seealso cref="EnableJoining"/>
+        /// <seealso cref="joiningEnabled"/>
         public void DisableJoining()
         {
             switch (m_JoinBehavior)
@@ -319,6 +337,7 @@ namespace UnityEngine.InputSystem
             m_AllowJoining = false;
         }
 
+        ////TODO
         /// <summary>
         /// Join a new player based on input on a UI element.
         /// </summary>
@@ -326,12 +345,12 @@ namespace UnityEngine.InputSystem
         /// This should be called directly from a UI callback such as <see cref="Button.onClick"/>. The device
         /// that the player joins with is taken from the device that was used to interact with the UI element.
         /// </remarks>
-        public void JoinPlayerFromUI()
+        internal void JoinPlayerFromUI()
         {
             if (!CheckIfPlayerCanJoin())
                 return;
 
-            //find used device
+            //find used device; InputSystemUIInputModule should probably make that available
 
             throw new NotImplementedException();
         }
@@ -363,21 +382,23 @@ namespace UnityEngine.InputSystem
             JoinPlayer(pairWithDevice: device);
         }
 
-        public void JoinPlayer(int playerIndex = -1, int splitScreenIndex = -1, string controlScheme = null, InputDevice pairWithDevice = null)
+        public PlayerInput JoinPlayer(int playerIndex = -1, int splitScreenIndex = -1, string controlScheme = null, InputDevice pairWithDevice = null)
         {
             if (!CheckIfPlayerCanJoin(playerIndex))
-                return;
+                return null;
 
-            PlayerInput.Instantiate(m_PlayerPrefab, playerIndex: playerIndex, splitScreenIndex: splitScreenIndex,
+            PlayerInput.s_DestroyIfDeviceSetupUnsuccessful = true;
+            return PlayerInput.Instantiate(m_PlayerPrefab, playerIndex: playerIndex, splitScreenIndex: splitScreenIndex,
                 controlScheme: controlScheme, pairWithDevice: pairWithDevice);
         }
 
-        public void JoinPlayer(int playerIndex = -1, int splitScreenIndex = -1, string controlScheme = null, params InputDevice[] pairWithDevices)
+        public PlayerInput JoinPlayer(int playerIndex = -1, int splitScreenIndex = -1, string controlScheme = null, params InputDevice[] pairWithDevices)
         {
             if (!CheckIfPlayerCanJoin(playerIndex))
-                return;
+                return null;
 
-            PlayerInput.Instantiate(m_PlayerPrefab, playerIndex: playerIndex, splitScreenIndex: splitScreenIndex,
+            PlayerInput.s_DestroyIfDeviceSetupUnsuccessful = true;
+            return PlayerInput.Instantiate(m_PlayerPrefab, playerIndex: playerIndex, splitScreenIndex: splitScreenIndex,
                 controlScheme: controlScheme, pairWithDevices: pairWithDevices);
         }
 
@@ -580,6 +601,20 @@ namespace UnityEngine.InputSystem
             if (actions == null)
                 return true;
 
+            // If the asset has control schemes, see if there's one that works with the device plus
+            // whatever unpaired devices we have left.
+            if (actions.controlSchemes.Count > 0)
+            {
+                using (var unpairedDevices = InputUser.GetUnpairedInputDevices())
+                {
+                    if (InputControlScheme.FindControlSchemeForDevices(unpairedDevices, actions.controlSchemes,
+                        mustIncludeDevice: device) == null)
+                        return false;
+                }
+                return true;
+            }
+
+            // Otherwise just check whether any of the maps has bindings usable with the device.
             foreach (var actionMap in actions.actionMaps)
                 if (actionMap.IsUsableWithDevice(device))
                     return true;